Refactoring/expanding upon protocol module
